put openshift-ansible compilation behind an env variable

+++ refresh.sh
@@ -1,7 +1,19 @@
 #!/bin/bash
 
-rm -rf /home/jrivera/projects/github/openshift/openshift-ansible/.tox
-tar -czf openshift-ansible.tgz -C /home/jrivera/projects/github/openshift openshift-ansible
-scp -rF ssh-config *ansible-inventory* openshift-ansible.tgz master:/vagrant/
-vagrant ssh master -c "cd /vagrant; mkdir -p repo/; rm -rf repo/openshift-ansible* repo/noarch openshift-ansible; tar --warning=no-timestamp -xzf openshift-ansible.tgz; cd openshift-ansible; tito build --test --rpm -o /vagrant/repo; createrepo /vagrant/repo; cd; sudo yum remove -y openshift-ansible*; sudo yum --disablerepo=* --enablerepo=openshift-local makecache; sudo yum install -y --enablerepo=epel-testing openshift-ansible*"
-rm -rf openshift-ansible.tgz
+scp -rF ssh-config *ansible-inventory* master:/vagrant/
+
+if [ $OPENSHIFT_ANSIBLE_DEVELOPER -eq 1 ]
+then
+        if  [ -n $OPENSHIFT_ANSIBLE_SOURCE_DIR ] && [ -d $OPENSHIFT_ANSIBLE_SOURCE_DIR ]
+        then
+                rm -rf ${OPENSHIFT_ANSIBLE_SOURCE_DIR}/.tox
+                tar -czf openshift-ansible.tgz -C ${OPENSHIFT_ANSIBLE_SOURCE_DIR}/../ openshift-ansible
+                scp -rF ssh-config openshift-ansible.tgz master:/vagrant/
+                vagrant ssh master -c "cd /vagrant; mkdir -p repo/; rm -rf repo/openshift-ansible* repo/noarch openshift-ansible; tar --warning=no-timestamp -xzf openshift-ansible.tgz; cd openshift-ansible; tito build --test --rpm -o /vagrant/repo; createrepo /vagrant/repo; cd; sudo yum remove -y openshift-ansible*; sudo yum --disablerepo=* --enablerepo=openshift-local makecache"
+                rm -rf openshift-ansible.tgz
+        else
+                echo "OPENSHIFT_ANSIBLE_SOURCE_DIR is not defined"
+        fi
+fi
+
+vagrant ssh master -c 'sudo yum install -y --enablerepo=epel-testing openshift-ansible*'
